requirements.txt
tensorflow-gpu
numpy
intervaltree
pytest
tensor2tensor
